KAFKA-6585: Consolidate duplicated logic on reset tools (#9255)
Reviewers: Navinder Pal Singh Brar <navinder_brar@yahoo.com>, Matthias J. Sax <matthias@confluent.io>
